+++ Src/EGame/Graphics/Vulkan/Pipeline.cpp
@@ -77,7 +77,7 @@ namespace eg::graphics_api::vk
 		{
 			VkSpecializationInfo* specInfo = linAllocator.New<VkSpecializationInfo>();
 			specInfo->dataSize = stageInfo.specConstantsDataSize;
-			specInfo->mapEntryCount = stageInfo.specConstants.size();
+			specInfo->mapEntryCount = (uint32_t)stageInfo.specConstants.size();
 			
 			void* specConstantsData = linAllocator.Allocate(stageInfo.specConstantsDataSize);
 			specInfo->pData = specConstantsData;
